Moved Binary Trees tests to BinaryTrees folder
